+++ src/app/components/turnos/punto-inicio/solicitud-turno-ventanilla/solicitud-turno-ventanilla.html
@@ -1,104 +1,2 @@
-<form class="plex-layout" #form="ngForm">
-
-    <fieldset>
-        <legend>Cargar Solicitud</legend>
-
-        <div class="row" *ngIf="modelo?.paciente?.id">
-            <div class="col-12">
-                <label>Paciente</label> {{ modelo.paciente | nombre }}
-                <span>Documento:
-                    <span *ngIf="modelo.paciente.documento !== ''">{{ modelo.paciente.documento }}</span>
-                    <span *ngIf="modelo.paciente.documento === ''" class="text-danger">Sin documento</span>
-                </span>
-                <span>Fecha de nacimiento
-                    <span *ngIf="modelo.paciente.fechaNacimiento !== null">{{ modelo.paciente.fechaNacimiento |
-                        date:'dd-MM-yyyy'
-                        }} ({{ modelo.paciente | edad }})</span>
-                    <span *ngIf="modelo.paciente.fechaNacimiento === null">Sin feacha de nacimiento</span>
-                </span>
-            </div>
-        </div>
-        <!-- Fecha Solicitud -->
-        <div class="row">
-            <!-- <div class="col-6" *ngIf=" !showSeleccionarPaciente"> -->
-            <div class="col-6">
-                <plex-datetime type="date" [(ngModel)]="modelo.solicitud.fecha" name="fechaSolicitud"
-                               label="Fecha de solicitud" [required]="true"></plex-datetime>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col d-flex">
-                <plex-bool class="align-self-center" [(ngModel)]="autocitado" label="Autocitado" name="autocitado">
-                </plex-bool>
-            </div>
-        </div>
-        <!-- Tipo Prestación Destino -->
-        <div class="row">
-            <div class="col">
-                <plex-select [(ngModel)]="modelo.solicitud.tipoPrestacion" name="tipoPrestacion"
-                             (getData)="loadTipoPrestaciones($event)" label="Tipo de Prestación Solicitada"
-                             placeholder="Tipos de Prestación solicitada" [required]="true"
-                             (change)="onSelectPrestacionOrigen()">
-                </plex-select>
-            </div>
-        </div>
-        <!-- Organización Origen -->
-        <div class="row">
-            <div class="col" *ngIf="!autocitado">
-                <plex-select [(ngModel)]="modelo.solicitud.organizacionOrigen" name="organizacionOrigen"
-                             [data]="dataOrganizacionesOrigen" label="Organización origen"
-                             placeholder="Seleccione la organización" labelField="nombre"
-                             (change)="onSelectOrganizacionOrigen()" [required]="true">
-                </plex-select>
-            </div>
-        </div>
-        <!-- Tipo Prestación Origen -->
-        <div class="row">
-            <div class="col" *ngIf="!autocitado">
-                <plex-select [(ngModel)]="prestacionOrigen" [data]="dataTipoPrestacionesOrigen" name="tipoPrestacion"
-                             (change)="onSelect()" label="Tipos de Prestación Origen"
-                             placeholder="Tipos de Prestación Origen" [required]="true">
-                </plex-select>
-            </div>
-        </div>
-        <!-- Equipo de Salud Origen -->
-        <div class="row">
-            <div class="col">
-                <plex-select [(ngModel)]="modelo.solicitud.profesionalOrigen" name="profesionalOrigen"
-                             (getData)="loadProfesionales($event)" label="Equipo de Salud solicitante"
-                             placeholder="Escriba el apellido del Equipo de Salud" labelField="apellido + ' ' + nombre"
-                             [required]="true">
-                </plex-select>
-            </div>
-        </div>
-        <!-- Equipo de Salud Destino -->
-        <div class="row">
-            <div class="col" *ngIf="!autocitado">
-                <plex-select *ngIf="!autocitado" [(ngModel)]="modelo.solicitud.profesional"
-                             label="Equipo de Salud destino" name="profesional" (getData)="loadProfesionales($event)"
-                             placeholder="Escriba el apellido de un Equipo de Salud"
-                             labelField="apellido + ' ' + nombre">
-                </plex-select>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-12">
-                <plex-text [(ngModel)]="motivo" label="Notas / Diagnóstico / Motivo" name="motivo" multiline="true">
-                </plex-text>
-                <!-- <label for="registros">Notas / Diagnóstico / Motivo</label>
-                <textarea class="form-control" [(ngModel)]="motivo" name="motivo"
-                    id="registros" rows="5"></textarea> -->
-            </div>
-        </div>
-        <div class="mt-3">
-            <plex-button class="float-left" type="danger" label="Cancelar" (click)="cancelar()">
-            </plex-button>
-            <plex-button class=" float-right" type="success" label="Guardar" (click)="guardarSolicitud($event)"
-                         title="Guardar Solicitud" validateForm="true">
-            </plex-button>
-            <!-- <plex-button type="success" label="Guardar" (click)="guardarSolicitud($event); listaSolicitudTurnoVentanilla.cargarSolicitudes()"
-                title="Guardar Solicitud" validateForm="true">
-            </plex-button> -->
-        </div>
-    </fieldset>
-</form>
+<form-nueva-solicitud [paciente]="paciente" size="sm" [tipoSolicitud]="'entrada'" (operacionFinalizada)="returnForm()">
+</form-nueva-solicitud>
